% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bake_bagofpatterns.R
\name{bake_bagofpatterns}
\alias{bake_bagofpatterns}
\title{Bake data into a Bag of Patterns}
\usage{
bake_bagofpatterns(bagofpatterns_obj, newdata = NULL)
}
\arguments{
\item{bagofpatterns_obj}{A model trained by \code{fit_bagofpatterns}}

\item{newdata}{An optional new data frame - if not supplied, the converted training data is returned.}
}
\description{
Convert a dataset of timeseries into a bag of patterns representation
}
\details{
This function uses a bag of patterns object trained on an example dataset
to return a histogram of 'words' representing the time series.
}
