% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FaceAll_TRAIN}
\alias{FaceAll_TRAIN}
\title{FaceAll Data Set (Training)}
\format{
A data frame with 560 rows and 132 variables:
\describe{
\item{attr\%}{The position of the face at a point - 131 variables}
\item{target}{The student, as a class}
...
}
}
\source{
\url{http://www.timeseriesclassification.com/description.php?Dataset=FaceAll}
}
\usage{
FaceAll_TRAIN
}
\description{
FaceAll was generated at UCR by Xiaopeng Xi. The faces belong to 14 grad students.
Each facial outline is mapped onto a one dimensional series. Sourced from
timeseriesclassification.com
}
\keyword{datasets}
