% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bagofpatterns.R
\name{fit_bagofpatterns}
\alias{fit_bagofpatterns}
\title{Fit Bag Of Patterns Histogram}
\usage{
fit_bagofpatterns(
  data,
  target = "target",
  window_size = 200,
  sparse_windows = FALSE,
  normalize = TRUE,
  alphabet_size = 4,
  word_size = 8,
  breakpoints = "quantiles",
  word_weighting = tm::weightTf,
  maximum_sparsity = NA,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame where each row is a time series, along with a column for class}

\item{target}{the name of the column where the class of each row is stored}

\item{window_size}{The size of the sliding windows as applied to the time series}

\item{sparse_windows}{a logical, indicating whether \code{sqrt(m)} random windows should be taken instead of all}

\item{normalize}{a logical, indicating whether each window should be z-normalized (\verb{(x - mean(x)/sd(x)})}

\item{alphabet_size}{the number of distinct letters to use in the compressed SAX representation}

\item{word_size}{the size of the 'words' generated out of the alphabet by SAX}

\item{breakpoints}{the method used to assign letters (see \code{seewave::SAX})}

\item{word_weighting}{The weighting function for the DTM/TDM (default is term-frequency, effectively unweighted)}

\item{maximum_sparsity}{A optional numeric for the maximal allowed sparsity in the range from bigger zero to smaller one.}

\item{verbose}{whether to print the progress of model creation.}
}
\description{
Fit a dictionary to a dataset using SAX to create a 'bag of patterns'
}
\details{
This function creates a dictionary to a dataset of time series using \code{seewave::SAX}
and returns an model that can be used by \code{bake_bagofpatterns} to convert that series
into a histogram of 'words'.
}
