% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_of_patterns_knn.R
\name{bagofpatterns_knn}
\alias{bagofpatterns_knn}
\title{Bag of Patterns (K-Nearest Neighbours)}
\usage{
bagofpatterns_knn(
  data,
  target = "target",
  k = 1,
  window_size = 200,
  alphabet_size = 4,
  PAA_number = 8,
  breakpoints = "quantiles",
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame where each row is a time series, along with a column for class}

\item{target}{the name of the column where the class of each row is stored}

\item{window_size}{The size of the sliding windows as applied to the time series}

\item{alphabet_size}{the number of distinct letters to use in the compressed SAX representation}

\item{PAA_number}{the size of the 'words' generated out of the alphabet by SAX}

\item{breakpoints}{the method used to assign letters (see \code{seewave::SAX})}

\item{verbose}{whether to print the progress of model creation.}
}
\description{
Fit a Bag of Patterns Model with KNN
}
\details{
This function fits a bag of patterns model by using Symbolic Aggregate
eXpressions (SAX) to convert a data frame of time series into a count table
of 'words' as generated by SAX generated by sliding windows over the series.
New examples are then fit by finding nearest neighbor examples with the converted
training data.
}
\examples{

}
